scripts/requirements.txt
loguru==0.7.2
pandas==2.2.2
polars==1.4.0
pyarrow==17.0.0
python-dotenv==1.0.1
rich==13.7.1
SQLAlchemy==2.0.30
pymysql==1.1.0